Moved `Upvalue` into `Compiler` as `UpvalueInfo`, to make way for the runtime `Upvalue` in 25.3.
